feat/add-preparation-days-endpoint

+++ Assert.Application/DTOs/Requests/UpsertMaxMinStayRequestDTO.cs
@@ -2,7 +2,7 @@
 
 public class UpsertMaxMinStayRequestDTO
 {
-    public int ListingRentId { get; set; }
+    public long ListingRentId { get; set; }
     public bool SetMinStay { get; set; }
     public int MinStay { get; set; }
     public bool SetMaxStay { get; set; }
